requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
pygit2
PyGithub
httpx
python-dotenv